fix: 스플래시 화면, 프로필 웰컴페이지 수정
