--- src/pages/login/login-end.tsx
@@ -1,7 +1,0 @@
-export const LoginEnd = () => {
-  return (
-    <>
-      <span>로그인 완료</span>
-    </>
-  );
-};
