+++ src/pages/splash/splash.tsx
@@ -1,11 +1,8 @@
-import { async } from "@firebase/util";
 import { motion } from "framer-motion";
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import { styled } from "styled-components";
 
-import SplashBackSVG from "@/assets/images/splash-back.svg";
-import SplashHandSVG from "@/assets/images/splash-hand.svg";
 import SplashLogoSVG from "@/assets/images/splash-logo.svg";
 import { colorTheme } from "@/style/color-theme";
 import getRefreshToken from "@/utils/token";
@@ -51,15 +48,6 @@ export const Splash = () => {
 
   return (
     <Wrapper>
-      {/* <BackgroundImg img={SplashBackSVG}>
-        <HandImg
-          src={SplashHandSVG}
-          animate={{
-            rotate: [0, -10, 10, -10, 10, 0],
-            transition: { duration: 2, repeat: Infinity },
-          }}
-        />
-      </BackgroundImg> */}
       <LogoImg
         src={SplashLogoSVG}
         initial={{ opacity: 0, scale: 0.6 }}
@@ -82,22 +70,4 @@ const Wrapper = styled.div`
 const LogoImg = styled(motion.img)`
   width: 15.06rem;
   height: 3.38rem;
-`;
-
-const BackgroundImg = styled.div<{ img: string }>`
-  width: 12.11rem;
-  height: 7.56rem;
-  position: relative;
-  background-image: url(${({ img }) => img});
-  background-position: center;
-  background-repeat: no-repeat;
-  background-size: contain;
-`;
-
-const HandImg = styled(motion.img)`
-  width: 6.167rem;
-  height: 6.06rem;
-  position: absolute;
-  left: 27%;
-  bottom: -16%;
 `;
